Clean up API visibility
