+++ kmmbridge-test/src/main/kotlin/co/touchlab/kmmbridge/test/TestArtifactManager.kt
@@ -1,9 +1,11 @@
 package co.touchlab.kmmbridge.test
 
+import co.touchlab.kmmbridge.KmmBridgeExtension
 import co.touchlab.kmmbridge.artifactmanager.ArtifactManager
 import org.gradle.api.Project
 import org.gradle.api.Task
 import org.gradle.api.tasks.TaskProvider
+import org.gradle.kotlin.dsl.getByType
 import java.io.File
 
 class TestArtifactManager : ArtifactManager {
@@ -20,3 +22,5 @@ class TestArtifactManager : ArtifactManager {
 
     override fun deployArtifact(task: Task, zipFilePath: File, version: String): String = url
 }
+
+internal val Project.kmmBridgeExtension get() = extensions.getByType<KmmBridgeExtension>()
