+++ kmmbridge/src/main/kotlin/co/touchlab/kmmbridge/Constants.kt
@@ -1,0 +1,4 @@
+package co.touchlab.kmmbridge
+
+internal const val TASK_GROUP_NAME = "kmmbridge"
+public const val EXTENSION_NAME = "kmmbridge"
