+++ kmmbridge/src/main/kotlin/co/touchlab/kmmbridge/dependencymanager/SpmDependencyManager.kt
@@ -15,16 +15,16 @@ package co.touchlab.kmmbridge.dependencymanager
 
 import co.touchlab.kmmbridge.TASK_GROUP_NAME
 import co.touchlab.kmmbridge.dsl.TargetPlatformDsl
-import co.touchlab.kmmbridge.findXCFrameworkAssembleTask
+import co.touchlab.kmmbridge.internal.findXCFrameworkAssembleTask
 import co.touchlab.kmmbridge.internal.domain.SwiftToolVersion
 import co.touchlab.kmmbridge.internal.domain.TargetPlatform
 import co.touchlab.kmmbridge.internal.domain.konanTarget
 import co.touchlab.kmmbridge.internal.domain.swiftPackagePlatformName
-import co.touchlab.kmmbridge.kmmBridgeExtension
-import co.touchlab.kmmbridge.kotlin
-import co.touchlab.kmmbridge.layoutBuildDir
-import co.touchlab.kmmbridge.urlFile
-import co.touchlab.kmmbridge.zipFilePath
+import co.touchlab.kmmbridge.internal.kmmBridgeExtension
+import co.touchlab.kmmbridge.internal.urlFile
+import co.touchlab.kmmbridge.internal.zipFilePath
+import co.touchlab.kmmbridge.internal.kotlin
+import co.touchlab.kmmbridge.internal.layoutBuildDir
 import org.gradle.api.Action
 import org.gradle.api.Project
 import org.gradle.api.Task
@@ -43,7 +43,7 @@ import java.nio.charset.Charset
 import java.util.*
 import javax.inject.Inject
 
-class SpmDependencyManager(
+internal class SpmDependencyManager(
     /**
      * Folder where the Package.swift file lives
      */
